break up CollectionFilter into constituent parts (e.g. files)
