+++ src/components/Firestore/CollectionFilter/CollectionFilter.module.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import '../common/color';
+@import '../../common/color';
 
 .panel {
   width: 360px; // taken from console
