+++ src/components/Firestore/CollectionFilter/index.ts
@@ -1,0 +1,2 @@
+import { CollectionFilter } from './CollectionFilter';
+export { CollectionFilter };
